+++ src/main/java/com/climatizacion/sistema_clima/dto/ServicioDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ServicioDTO {
-    private Integer idServicio;
+    private Long idServicio;
     private String nombre;
     private String descripcion;
     private Double precio;
