+++ src/main/java/com/climatizacion/sistema_clima/entities/SolicitudServicioEntity.java
@@ -14,7 +14,7 @@ public class SolicitudServicioEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idSolicitud;
+    private Long idSolicitud;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_cliente", nullable = false)
