+++ src/main/java/com/climatizacion/sistema_clima/repository/DetallePedidoRepository.java
@@ -3,5 +3,5 @@ package com.climatizacion.sistema_clima.repository;
 import com.climatizacion.sistema_clima.entities.DetallePedidoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetallePedidoRepository extends JpaRepository<DetallePedidoEntity, Integer> {
+public interface DetallePedidoRepository extends JpaRepository<DetallePedidoEntity, Long> {
 }
