+++ src/main/java/com/climatizacion/sistema_clima/repository/PedidoRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PedidoRepository extends JpaRepository<PedidoEntity, Integer> {
-    List<PedidoEntity> findByIdUsuario(Integer idUsuario);
+public interface PedidoRepository extends JpaRepository<PedidoEntity, Long> {
+    List<PedidoEntity> findByIdUsuario(Long idUsuario);
     long countByEstadoIn(List<String> estados);
-    long countByIdUsuarioAndEstadoIn(Integer idUsuario, List<String> estados);
+    long countByIdUsuarioAndEstadoIn(Long idUsuario, List<String> estados);
 }
