+++ src/main/java/com/climatizacion/sistema_clima/repository/ServicioRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ServicioRepository extends JpaRepository<ServicioEntity, Integer> {
+public interface ServicioRepository extends JpaRepository<ServicioEntity, Long> {
 }
